[mk] Try to detect stale working directories when possible.
configure and build targets explicitely cd to the working directory. It
is possible to detect a failure a this level and issue an error message
if the working directories don't exist. Usually this is because of a
stale working directory.

The same error is more difficult to detect for other targets, but this
is less critical since error do not normally happen for other targets
than configure or build.