Fix 'pathsearch' macro, and define a new 'prefixsearch' one.
If pathsearch is given a list of files, it returns the full path of all
files as a list, instead of the full path of the first file found.
Fix pathsearch invocation in python.mk accordingly.

The prefixsearch macro is similar, except that it returns only the prefix
of the files that are found (i.e. for a path f, it returns the string s
such as the file "s/f" exist).