Add a new dependency computation system that can deal with system packages.
A package can now define a "depend.mk" file that specify how this package
can be replaced by a system package (i.e. a package that is already
provided by the system).

The depend.mk must specify a list of system files to look for in order to
detect an already installed system package. User can indicate their
preference by defining, in robotpkg.conf and for each package <pkg>, a
variable PREFER.<pkg> that can be set to either "system" or "robotpkg".