[archivers/zlib] Import zlib 1.2.3 - defaults to system package
zlib is a general purpose data compression library. The robotpkg_* admin
tools now require it. In order to avoid the "robotpkg Linux distribution"
syndrome, this package defaults to the system zlib package. A robotpkg
fallback version is still provided because a/ it is small and b/ that can
help to automate the bootstrap process for people not familiar with their
OS distribution ...