[archivers/bzip2] Upgrade to bzip2 1.0.5 - now default to system package
The recent robotpkg_* admin tools now require bzip2. As for the zlib
package, this one now defaults to the system package of bzip2. However, a
fallback is provided if, for some reason, the system bzip2 is not
available or cannot be used.

Changes since bzip2 1.0.4:
Security fix only.  Fixes CERT-FI 20469 as it applies to bzip2.