[mk] Change the way options are handled.
Supported options can now automatically be added to packages that
require other packages (like for a genom module that requires genom and
inherit from the "tcl" option).

The separate "option.mk" file in every package is no longer required and
the options are processed in the main stream Makefile.

A package can define the following variables:

PKG_SUPPORTED_OPTIONS, for the list of supported options.

PKG_OPTION_DESCR.<opt> for a text description of the option <opt>.

PKG_OPTION_SET.<opt>   contains a script that is $(eval ) if the <opt>
                       option is set by the user.

PKG_OPTION_UNSET.<opt> is the same as above but is eval'ed when <opt>
                       is not set.

The ${PKG_OPTION_VAR} user-settable option variable now defaults to
PKG_OPTIONS.${PKGBASE}, which is the most common case and provides some
homogeneity among packages.