[audio/festival] Import festival-1.96beta
Festival offers a general framework for building speech synthesis
systems. As a whole it offers full text to speech through a number APIs:
from shell level, though a Scheme command interpreter, as a C++ library,
from Java, and an Emacs interface.

Festival is imported into robotpkg because the source distribution is a
little bit messy and systems packages might exhibit big differences
(e.g. the NetBSD package does not provide the C++ headers and the client
library). However, we still default to the system package in robotpkg.

This source distribution is the one of the festvox project from CMU.