[pkgtools/libtool] Make pkgtools/libtool really optional. Removed USE_LIBTOOL
You can now safely `robotpkg_delete libtool`.

The USE_LIBTOOL variable is obsoleted. It is replaced by the inclusion
of the regular pkgtools/libtool/depend.mk depend file. All packages that
where using this variable have been updated.

libtool package now defaults to a system package, but can optionaly be
compiled by robotpkg if the system does not provide one that is usable.

The libtool depend.mk includes some of the logic that was previously
disseminated into the mk/* files, especially the post-configure target
that makes sure that the proper libtool script is used when compiling the
packages.