[pkgtools/libtool] Correctly replace all libtool files in pre-configure
When used in conjunction with autoconf, libtool must be generated by
autoconf. In order to make sure that autoconf will generate a libtool
script that is using the version found by robotpkg, one has to update
config.sub, config.guess and ltmain.sh (that's what libtoolize does).

Specifying a dependency on pkgtool/libtool in some package will trigger
this replacement procedure. If this is not wanted, the package should not
depend on libtool. In order to customize the replacement process, the
package can set LIBTOOL_OVERRIDE to a list of files to be replaced and
define OVERRIDE_PATH.<file> for each file in LIBTOOL_OVERRIDE (see
pkgtools/libtool/depend.mk).