[pkgtools/pkg-config] Make pkg-config optional, default to the system pkg.
You can now safely `robotpkg_delete pkg-config`.

The USE_PKG_CONFIG variable is obsoleted. It is replaced by the inclusion
of the regular pkgtools/pkg-config/depend.mk depend file.

pkg-config package now defaults to a system package, but can optionaly be
compiled by robotpkg if the system does not provide one that is usable.

The pkg-config depend.mk includes some of the logic that was previously
disseminated into the mk/* files, especially the pre-configure target
that makes sure that rpaths are properly used.