[mk] Cleanup resolve.mk and simplify the PKG_CONFIG_PATH definition.
Make use of new macros to simplify the definition of the PKG_CONFIG_PATH
variable that must find all system packages that are used.

XXX This should really be in pkg-config/depend.mk but it's not possible
to have the complete list of dependencies at the time the depend.mk
file is parsed.