[image/coriander] Import coriander 2.0.0
Coriander is the Linux graphical user interface (GUI) for controlling a
Digital Camera through the IEEE1394 bus (aka FireWire, or iLink).

Coriander is full featured and besides changing the parameters of
the camera it will also let you record video, send images to an FTP site,
convert the video to a V4L stream,... A live display is of course
provided too. Best of all, Coriander will work with any camera that is
compatible with the IIDC specifications (also known as DCAM specs). This
includes most 1394 webcams and a majority of industrial or scientific
cameras too.