[cross/i386-mingw32] Import mingw 3.14
MinGW is a collection of header files and import libraries that allow
one to use GCC and produce native Windows32 programs that do not rely
on any 3rd-party DLLs. See http://www.mingw.org/

This package is heavily inspired from the same package in pkgsrc, but
uses more recent versions of mingw runtime (3.2) and gcc (3.4.5).