[mk] Make the prefixsearch script smarter in looking for system files.
. Added a csh-like brace substitution for system files specs (e.g. a{b,c}d
expands to 'abd acd') thanks to a simple awk script. This make it possible to
specify alternative locations for some files (e.g. tcl that installs some files
in lib/tcl-x.x on ubuntu and directly in lib on other systems).

. When a system file spec contains shell wildcards, iterate over all the words
of the expansion instead of only checking the first one. This increases chances
that one file will match the spec (if e.g. several tcl version are installed
this will make it possible to eventually find the requested one).

. Make sure to replace all dashes (-) by dots (.) in the system files versions
strings so that the `pkg_admin pmatch` command do not get confused. pkg_admin
relies on the convention that a package is named after the <pkg>-<version>
template, with no dash in the version part. This should fix the flex detection
problem on ubuntu where the `flex -V` returns something like 2.5.x-y.z (y.z is
the ubuntu patch level).