[lang/gcc] Fine-grain tune gcc4 selection
Add meta-packages for gcc4-{c,c++,fortran} that actually install their
gcc42-{c,c++,fortran} sibling package. These gcc4 packages are normally not
needed but allow the user to select gcc4 from system or robotpkg in a way
not dependant from the gcc42 selection.

This solves cleanly a problem on systems that ships with gcc-4.1 that is fine
for most packages but the few ones that require gcc-4.2 explicitely.