[mk] Implement the checkout target.
`make checkout` performs a repository checkout if one has been defined in the
variable MASTER_REPOSITORY. The checkout target replaces the regular extract
target once it has been invoked and until the next `make clean`. At the moment
the following tools are supported for checkout: git, cvs.

MASTER_REPOSITORY must be of the following format for git:
  git <url>
and of the following format for cvs:
  cvs <root> <module>

The checkout is performed in ${WRKSRC}.

A package that is checkouted rather that extracted has its version number
suffixed by 'rc' plus the current date in the format y.mo.d.h.mi.s

Regular targets like configure, build, install, replace work as usual.
If a checkout is present in ${WKRSRC}, it prevents the automatic execution of
the targets fetch, checksum and patch, but these are still available manually.