[mk] Add su-target-% for running make install as root
make install and make clean now support just-in-time sudo feature.
This is useful for building software as a regular user and installing it with
root privileges.

The behaviour is controlled with the following variables to be put in
robotpkg.conf:

MAKE_SUDO_INSTALL: whether to use just-in-time sudo feature for make
install. Possible values: yes, no. Default: no. This one has to be changed to
use to jit-sudo feature.

SU_CMD: the sudo command. Defaults to "sudo -u root sh -c". Possible values:
any command that accepts commands like "sh -c". Does normally not need to be
changed.

SU_USER: the user to su to if the current user is different. Defaults to
"root", possibles values: any user name or id.