[hardware/sahand] Restrict sources and binaries and fix some details.
Use the NO_PUBLIC_* variables to restrict the use of this package.

If sources are not restricted, the daily 'mirror-distfiles' target will mirror
the distfile on softs.laas.fr/openrobots/robotpkg/distfiles, as a backup target
if MASTER_SITE is not available. If binaries are not restricted, binary
packages might be built if we did have automatic binary packages generation
(but we don't ;).

Also fixed a few details:
. PREFIX is the actual prefix of this package (might be different from
  ROBOTPGK_BASE).

. NO_CONFIGURE can be set to disable the configure target (turns it into a
  no-op).

. Use SU_CMD instead of (undefined) SUDO. This will systematically sudo the
  install of the kernel module, which not the same as before, but is more
  consistent with the rest. SU_CMD defaults to "sudo -u ${SU_USER} sh -c" and
  SU_USER defaults to root.