[mk] Teach robotpkg about cmake and source configuration with cmake.
Include mk/sysdep/cmake.mk in a package and it will use cmake to configure the
build tree. Usual variables are used, namely:

CONFIGURE_DIRS, a list of directories in which to invoke cmake, relative to
WRKSRC or absolute.

CMAKE_ARGS, a list of extra parameters to be passed to cmake.