[interfaces/jrl-mal] Fix PLIST - remove dynamicaly generated doxygen files.
When using sysdep/doxygen.mk, a DOXYGEN_PLIST_DIR variable is available and
should contain the directories (relative to PREFIX) where doxygen will install
its documentation.

The files in DOXYGEN_PLIST_DIR will be added dynamically to the PLIST, since
their name highly depend on the version of doxygen and we don't control which
doxygen is used from robotpkg.