[math/lapack] Fix linkage libtool tag name
If we link a gfortran library with libtool --tag=CC, libtool replaces the
linker with gcc and the dependency on libgfortran is not registered.

We have to use --tag=F77, but this currently breaks the compilation when using
gfortran from robotpkg. So building with PREFER.gcc4-fortran = robotpkg is not
supported at the moment.