[robots/hrp2-14] Add missing c++ in USE_LANGUAGE
USE_LANGUAGE must contain the list of languages used by the package.

For an obscure reason, cmake doesn't like to have CC=/usr/bin/gcc and CXX=g++,
which is what happens if USE_LANGUAGE isn't told to look for a c++
compiler. This is probably a good idea but the problem is that in that case,
cmake forget about all defines on the command line (e.g. the prefix) and
continue without any further errors... This breaks the installation since the
default prefix is not what the user want.