[bootstrap] Create prefix directory with appropriate permissions.
The bootstrap script now checks that <prefix> exists and is rwx by the
user invoking ./bootstrap.

If <prefix> does not exist, it is created with mkdir -p, falling back to
sudo mkdir -p in case of insufficient permissions. If something fails,
the error is printed and the user is asked for manual creation and
bootstrap stops.

If <prefix> is not rw by the user, a message is also printed and
bootstrap stops.