[math/blas] Replace "libtoolization" by "cmakeification"
Blas does not come with a portable Makefile, and the previous
"libtoolization" had problems dealing with robotpkg Fortran compiler
selection. Since building a clean libtool for this package would require
building a configure.ac script and include all the needed files, it's
easier to write a cmakefile.

While here, bumped the (unknown) blas version to the same as lapack (like
all the linux distributions do).