[mk/update] Exchange an old grep+awk with a brand new sed oneliner (less CO2)
The packages that have to be reinstalled during make update were computed
using a grep on pkg_delete -n plus an awk for printing the lines in reverse
order (smallest dependency first). All this can be done with a variation
of the 'G;h' sed oneliner.

robotpkg preserves the Earth: the less forks, the less carbon dioxyde.