[mk/sysdep] Give user a hint as to which package to install when one is missing
For system packages, the variable SYSTEM_PKG.${OPSYS}-${OPSUBSYS}.<pkg>
can be set to the name of the standard system package for the current
OPSYS and distribution (for Linux).

This variable is printed when the corresponding system package cannot be
found. I hope this will help the users that send me e-mails,
complaining their xxx package is installed and robotpkg does not work,
while what is really needed is xxx-devel (for Linux).

While here, print the most important lines of the error message in
*bold*, so that the users that don't read error messages get a chance to
actually read them before sending e-mails ;)