[mk/checksum] Make sure robotpkg.conf wasn't modified in the middle of a make
A new target `check-configuration-file' either create a checksum of the
current robotpkg.conf, or checks it against a previously saved checksum. This
guarantees that the configuration file did not change between two make
invocations (which could lead to e.g. dependencies problems).

If the checksum doesn't match, the user is asked to make clean before
continuing.