[mk/internal] Improve tests for interactive stdin/stdout
Inclusion of a generated file for testing -t stdin and stdout didn't work
well in some circumstances. Take a more conventional approach by creating a
phony target that touches temporary files and define 'bold' macros to look at
this file at run-rime.

This is a bit less efficient but those macros are mostly invoked for error
messages so this doesn't hurt.