[mk/depends] fix prefixsearch if the command for obtaining a version has spaces
When the prefixsearch script looks for the version of a system tool and the
command for obtaining the version contains a quoted string with spaces in it,
an extra 'eval' must be added before executing the command.

This fixes a problem with perl for which a possible command to get the version
is cmd='perl -le "print $^V"'. Doing v=`$cmd` doesn't work because cmd is
expanded only once and the resulting string then split into args directly (the
internal quotes aren't interpreted). Doing `eval $cmd` is correct because the
quotes inside $cmd are parsed again by the shell after the first expansion.

Hopefully, if the command has no special substrings the extra eval doesn't
hurt, except if there are special shell characters (e.g. dollars) in the
command. But there are no such commands yet.