[mk] include files for toplevel targets only when required
Since make fails when the robotpkg_info tool isn't working or installed, it was
not possible to do `make index` on a fresh checkout. Circumvent this by making
the inclusion of all toplevel targets conditionned to the target requested on
the command line.

In addition, most of the inter-dependencies between files are explicited and
there should be less constraints on the subtle ordering of the include
directives in the main robotpkg.mk file, making maintenance easier.