[mk/depends] Move the system dependencies resolution engine to its own target
Instead of writing complex Make code to check that system packages are present
and compute their prefixes, a dedicated target has been created (with less
complex shell code ;). It is invoked during `bootstrap-depends' and creates a
file with prefix information as reported by `prefixsearch.sh'. This file is
then included to make the PREFIX.<pkg> variable available in packages
makefiles. Information is guaranteed to be correct only after the `barrier'
target, but this is exactly what is needed (and before there was no guaranty at
all because of the file inclusion order problem...).

The main benefit is that system dependencies can now be defined anywhere in the
makefiles: the list will always be correct since it is evaluated in a target
context, after all Makefiles have been parsed.

The second benefit is that dependencies are not checked when this is not
necessary. This finally make the `index' target work out-of-the-box, even
before bootstrapping robotpkg, so that new users can generate html package
files before compiling anything.