[mk/options] Always define the `debug' option; let compilers choose the action
`debug' is a common option, so define it unconditionnaly in
robotpkg.options.mk (provided there is a `build' step). However, the flags
corresponding to this option are still defined by the compilers (gcc only
atm).

This makes the robotpkg.options.mk file independent of compiler-vars and it can
now be included unconditionnaly, so that the full set of options are always
available to Makefiles. Since the dependencies of a package can be changed by
an option this is more robust to do like this.