[mk/depends] Better handling of shell quoting in prefixsearch.sh
In order to cope with the builtin echo in dash and its backslash expansion (I
personnaly would say 'buggy expansion'), choose a different strategy to call
the script looking for the system packages.

The main point is to completely avoid using 'echo' when dealing with quoted
parameters. This was done by changing the calling strategy of the prefixsearch
script (at the cost of longer command lines) and also using the -r parameter of
read for parsing the file:spec:cmd string into individual variables.

Now, all undesired expansions by the shell are avoided.