[hardware/pa10-libs] Fix ABI requirement in depend.mk
DEPEND_ABI.pa10-libs defines the minimum version compatible with all the
packages that depend on it. It's not necessary to bump it each time.
Furthermore, the installed .pc is still set to version 1.1 ...

Also, simplified the SYSTEM_SEARCH variable: it doesn't have to contain a lot
of files. Just a few, typical files are required. Of course putting too much
files doesn't hurt, except for the carbon footprint of robotpkg ;)

Added a check on the .pc version, for robotpkg to find it in the case pa10-libs
would be installed outside the robotpkg tree.