[mk/plist] Change the timestamp file for print-PLIST to _COOKIE.build
Using the extraction date is not very robust when working in a `checkout'
mode. Using the installation start date doesn't work either when working on a
NFS filesystem, because of the timestamp resolution.

So use the time of last build. Use a `make rebuild' to update this timestamp if
the print-PLIST outputs too many files.