[mk/compiler] Fix a subtle bug when _GCC_REQUIRED was used in MAKEOVERRIDES
This commit fixes a problem with packages depending on viam-genom and getting
spurious error messages regarding viam-libs opencv build option.

_GCC_REQUIRED is the strictest constraint matching all GCC_REQUIRED statements.

Computing this value is expensive (calls `pkg_admin pmatch' several times), so
it was computed once and exported in sub-makes via the MAKEOVERRIDES variable
with appropriate quoting since the value typically contains >= or <=
strings. After 2 or more sub-make levels, this variable happened to be
underquoted, causing various errors in targets commands that used it
implicitely via MAKEOVERRIDE.

Anyway, this approach was wrong since the GCC_REQUIRED for a package is not the
same as for another package so it shouldn't be exported. Since most packages
don't define constraints on gcc, it's simpler to compute _GCC_REQUIRED each
time but only when GCC_REQUIRED is defined.