[mk/update] Better handling of `make clean' errors after an `update'
Rename NOCLEAN to UPDATE_CLEAN for improved clarity (double negations are
unbearable). This variable was undocumented and used at only one place so
renaming is cheap.

`clean-update' defaults to cleaning for regular packages, and to not cleaning
for checkouted packages. Suggested by O. Stasse in 57e9610. This can be
overriden in robotpkg.conf by setting
UPDATE_CLEAN= yes|no

If for some reason `make clean' fails after an update, print the error message
but don't fail immediately. After cleaning all directories, print a summary of
the files leftover. This can happen if we reinstalled packages depending on the
one we are updating, or if the user has overriden the default UPDATE_CLEAN
settings.