[mk/depends] Make prefixsearch.sh output the version of the system pkg found
For packages that define version checks in SYSTEM_SEARCH, prefixsearch will
print first the actual version found, then the Make variables for later
inclusion by depends.mk

If the file descriptor 3 is open, the Make variables are printed to this fd, to
stdout otherwise. When prefixsearch is called from depend.mk, we get the
package version in a shell variable from stdout and still create the sysdep
file at the same time by redirecting fd 3. When prefixsearch is called
with no redirections (for debugging), it behaves intuitively by printing
everything to stdout.