[mk/options] Fix common `debug' compile option definition.
The `debug' option was broken since the reordering of the file inclusion in
robotpkg.mk, because of a chicken-egg problem: robotpkg.options.mk needs
compiler-vars.mk to have the debug option definition and compiler-vars.mk as
well as other mk files need options to be processed before they are included.

Fix by using the 'deferred' variable evaluation of Make and define the debug
option through generic and compiler-independant intermediate variables. Actual
compilers can set the flags in those variables later, when they are included.