[image/artoolkitplus] Import ARToolKitPlus 2.1.1
ARToolKit is a software library that can be used to calculate camera position
and orientation relative to physical markers in real time. This enables the
easy development of a wide range of Augmented Reality applications.

The Makefile and patches are rather horrid, due to ARToolKit's clever
Makefiles, smart usage of `_*' reserved symbols and revolutionary template
writing style. Tested on Fedora 9, NetBSD 5.