[mk/plist] Make print-PLIST depend on `install' and run after the barrier
print-PLIST requires an freshly installed, so better depend on
`install' so that users don't get confused if they didn't install the
package beforehand.

Also, print-PLIST might require some variables that are only defined
as a result of the depends target, like all the sysdep-generated
variables. This is possible only if print-PLIST is invoked after the
barrier.

Since running after the barrier triggers extra verbose messages,
print-PLIST doesn't output its guess to stdout but to PLIST.guess by
default. Thus, the usage is now:
      make print-PLIST; cat PLIST.guess | brain > PLIST

[ Note that the brain utility is not included in robotpkg (for now). ]