[mk/sysdep] Improve python dependency by using the regular sysdep scheme.
Python dependency used to be defined in tools/python.mk by using an
ad-hoc Makefile and the old TOOLS_* variables from pkgsrc. It's easier
and now more robust to use the sysdep makefiles for that.
Two variables are defined : PYTHON and PYTHON_SITELIB. PYTHON is added
to the list of ignored environment variables.

As a result, update the openrtm-aist package that was using the old
scheme.