[pkgtools] Fix pax and ftp for Fedora 11
stdio.h in Fedora 11 now define the getline() function. It used to be
protected by an ifdef __GNU_SOURCE and is now protected by a
__USE_XOPEN2K8 which is activated by default.

pax and ftp defined their own getline function, not compatible with
Linux one, so we just prefix the pax and ftp version by pax_ and ftp_
respectively.

Keep track of this change as a separate patch so that upgrade is easier.