Add the incomplete documentation in order to have at least some starting point
The generated web pages mention the documentation but it was not
commited yet. This is especially annoying when a user find a package
webpage from google and has no way to find out how to use/install it.

Even though the current doc is far from complete, it documents at least
how to download and install something.