[archivers/libarchive] Make sure autoconf/automake or autoheader are not called
Disable the rules in Makefile.in that invoke autoconf et al.

This is useful for the "in place" build of this package: since the
sources are copied in the working directory, the timestamps might
change and configure.ac might (wrongly) appear newer than the
Makefile.in files. In such a case, the build will fail if autoconf is
not installed and adding an otherwise useless dependency on autoconf
should be avoided.

The diff is stored in files/patch-ab to ease the upgrade to a new
version of this package.