[architecture/yarp] Update yarp to 2.2.4
From announce mail :

There have been plenty of small improvements to wrappers for Python,
Java, etc.  Ruby is now supported.  YARP's method for searching for the
name server was made more conservative, so that searches only occur on
explicit request.  A new experimental name server is included in this
release.  It is optional, and not compiled by default. Compile it by
turning on the CREATE_YARPSERVER3 option in CMake. Various bugs have
been fixed, and optimizations made, see the ChangeLog for details.

Note that CREATE_YARPSERVER3 is automaticly set in this package.