[mk] Temporary workaround for resolving all dependencies in extract/checkout
extract and checkout targets don't require all the package
dependencies, but they are still run after the barrier. This can lead
to partially resolved dependencies (e.g. compilers are not checked),
making further `make build' fail.

A proper fix would be to have the extract/checkout (and patch) targets
run before the barrier and split the depend target into a bootstrap
and real depend targets.

The temporary workaround makes the extract/checkout target depend on
all the full dependencies of the package, which should fix the proble
for most usages.

This problem was reported by Thibault Kruse.