[mk/internal] Make barrier target not depend on 'depends'
While here, introduce a new macro to ease the writing of top-level
targets that use the barrier.

The idea behind this change is to allow 'extract' and 'patch' to work
even without all dependencies resolved (they actually require only
bootstrap-depends). This is useful to quickly browse a package's source
without installing all its dependencies.

This commit currently breaks most of the top-level targets. This will
be fixed in upcoming commits.