[mk] Fix PKG_DBDIR not having a default value (only defined by robotpkg.conf).
Make sure that robotpkg.default.conf defines the default PKG_DBDIR, in
case the user robotpkg.conf does not define it.

While here, make bootstrap not write the value of this variable  in
sample robotpkg.conf if it is equal to the default.

Update the default value to $LOCALBASE/var/db/robotpkg instead of
var/db/pkg. This will not break existing installations since the
value used to be defined systematically in robotpkg.conf, and was not
defined elsewhere (so users dit not delete it for sure).