[mk] Remove useless lines from sample robotpkg.conf produced by bootstrap
PKG_TOOLS_BIN needs not being defined in sample robotpkg.conf, since
the value cannot be configured by the bootstrap script. Just let the
default value in robotpkg.default.conf be used.

MAKE should really not be defined either. I remember this was
introduced when gmake-3.81 was not present in all linux distributions
(gmake-3.80 segfaults in robotpkg) but this should be OK now.