[math/fsqp] Fix FETCH_MESSAGE information
This package is actually a repackaged version of the original software
from AEM Design Inc. Mention ${MAINTAINER} as the contact for sources
instead of AEM Design.

This situation will be clarified soon by removing this package in
favor of the upcoming new fsqp package resulting from the split of
optimization/roboptim-core-cfsqp-plugin.