[mk/fetch] Add a boostrap depend on tnftp is some distfiles are missing
Now that pkg_install-20091115 do not depend on pkgtools/tnftp, it must
be pulled as a boostrap dependency by the fetch target. It is not
pulled if all distfiles are found, though.

XXX This requires a bit more work: since the ftp program is
user-configurable in mk.conf, tnftp should not be pulled blindly.