[pkgtools/pkg_install] Force pkgdb rebuild when upgrading from older version
In some cases, the new pkg_admin tool is not able to deal with a old
version of the packages database. This is probably due to differences
in the old db code and the new one included in libnbcompat.

To be on the safe side, systematically rebuild the database when
upgrading (i.e. the installed version is different from the current
version).