[mk/install] Fix a corner case problem with stale work directories
The install targets checks for stale working directories by comparing
the extracted version and the package version.

The action in case of inconsistencies was to clean and rebuild
automatically, but this could lead to problems with installation log
file. Better stop with an error, print an explanatory message and let
users clean themselves.