[mk/compiler] Define gcc42 as valid compiler for ROBOTPKG_COMPILER
If ROBOTPKG_COMPILER defines gcc42 (instead of gcc), the package in
lang/gcc42 is used for searching for a C/C++/Fortran compiler. The
usual variable PREFER.gcc42 (or PREFER.gcc42-{c,c++,fortran}) can be
set to system (default) or robotpkg to choose between system or
robotpkg pacakge. If 'system' is choosen, the compiler version must be
compatible with >=4.2<4.3.

This makes the selection of the gcc-4.2.4 in lang/gcc42 less cumbersome
and more "standard".

XXX there is currently a problem when using the robotpkg version of
gcc42 and some autoconf (+libtool) scripts and packages using only C in
USE_LANGUAGES. Such autoconf scripts will insist hard for finding a C++
preprocessor (even though they don't use it). The C++ preprocessor
found by the gcc42 package will not work since gcc42-cpp is compiled
only for C. (This paragraph is absolutely not clear).