[lang/python3] Added a package for Python 3.
Python 3.0 (a.k.a. "Python 3000" or "Py3k") is a new version of the language that is
incompatible with the 2.x line of releases. The language is mostly the same, but many
details, especially how built-in objects like dictionaries and strings work, have changed
considerably, and a lot of deprecated features have finally been removed. Also, the
standard library has been reorganized in a few prominent places.