[mk/configure] Fix cxxpp fatal errors in configure scripts from autoconf 2.64
Autoconf-2.64 generates slightly different code and we have to adapt
the regexp that disable cxxcpp fatal errors in packages not using it.
This will fix (at least) localization/insitu configure target.

This is of course a hack, but if it would fail too often it can always
be avoided by declaring the use of a (useless) C++ compiler in the
failing packages.