[image/libdc1394-1] Fix PKGNAME so that this pkg can coexist with libdc1394-2
libdc1934-1 and -2 can be installed simultaneously since they don't
share headers or libraries. Fix PKGNAME of version 1 so that it does
not conflict with that of version 2.

Disabled the dc193_vloopback example program because this one is also
installed by version 2.

While here, fixed other glitches such as old-dependency style on
libraw and regen PLIST.