[hardware/pa10-kmod] Split install and use an explicit sudo to install kmod
. Fixed broken PLIST (files starting with / are still relative to ${PREFIX})
. Patched the makefile so that it installs the kernel module under ${PREFIX}
. Added post-install that invokes sudo to copy the kmod to its final
  destination

Todo: the kmod in /lib/modules will never be removed (but the copy
under ${PREFIX} will). This should be handled by a DEINSTALL script,
which is not implemented yet.

It would also be possible to use a hardcoded MAKE_SUDO_INSTALL=yes
here, but it is a bit aggressive and the PLIST will not contain files
outside PREFIX anyway, so this will not improve anything regarding the
todo above.