[image/visp] Import visp-2.4.4
Based on initial package submission in wip/visp contributed by Thomas
Moulard.

ViSP (ViSP states for "Visual Servoing Platform") is a modular software
that allows fast development of visual servoing applications. The
platform (in fact a library) is implemented in C++ and is available
under Linux, Windows and OSX. ViSP is developped and maintained by the
INRIA Lagadic team located at INRIA Rennes.

This package provides the following patches:
 - patch-aa: Disable hardcoded -rpath option in CMakefile, this is
             cmake's job and generates command line weirdness anyway.
 - patch-ab:
 - patch-ac: Add missing include cstdlib for exit(2) in a few example files
 - patch-ad: Install cmake templates in share/cmake instead of lib.

A lot of compile-time options are available, see make show-options for
detail. ffmeg support is disabled as it is not compiling for some
strange reasons that seem unrelated to ffmpeg, at least on F11.