[bootstrap] Improve the bootstrapping of an existing installation.
It is sometimes necessary to (re)bootstrap an existing installation,
for instance when the robotpkg_* tools do not work after a system
upgrade.

In this case, bootstrap will abort if there is an existing
robotpkg.conf. This is necessary because the existing robotpkg.conf
might contain customizations that might confuse the default bootstrap
strategy. The user is asked to move the robotpkg.conf file away and
merge the settings manually after bootstrap.

bootstrap will also remove all pkg_install entries in pkgdbdir so that
the reinstallation of the complete robotpkg_ tool chain will be
triggered by robotpkg dependencies upon next package update.