[mk] Let the default SYSTEM_PREFIX be computed from the user PATH variable
SYSTEM_PREFIX is the ordered list of paths to be searched for system packages.

In order to robustify the search for 'simple' dependencies (simple meaning for
instance a single program like tar or unzip) that might reside in the user's
PATH but not in system's directories, it is convenient to build the default
SYSTEM_PREFIX out of user's PATH.

The default value is now made of the following values, in this order:
 - all ${PATH} components, with /bin and /sbin suffixes removed (because in the
   general case we are really looking for a prefix containing a bunch of files,
   not the path to a single file). PATH components without /{,s}bin suffixes
   are left unchanged, allowing for unusual path components on e.g. Darwin and
   its Frameworks.
 - /usr
 - /usr/local
 - /opt/local on Darwin.

Duplicates are removed, keeping only the first instance (in the place where it
first appears).

Any value defined in robotpkg.conf will completely override this default
value.