[mk/sysdep] Look for pdflatex in ${SYSTEM_PREFIX}/{bin/,}pdflatex
Make the bin/ part of SYSTEM_SEARCH.pdflatex optional. This allows pdflatex to
reside directly in a component of SYSTEM_PREFIX so that installations of
pdflatex that don't come from MacPorts and put pdflatex in /usr/texbin/pdflatex
can still be used (provided that /usr/texbin is in user's PATH or SYSTEM_PREFIX
explicitely contains /usr/texbin in robotpkg.conf).